% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{AUC}
\alias{AUC}
\title{AUC}
\usage{
AUC(predizioni, testpharmat, vectorAUC, name)
}
\arguments{
\item{predizioni}{matrix of predictions}

\item{testpharmat}{matrix of test for the side effects}

\item{vectorAUC}{empty vector where the AUC values will be saved}

\item{name}{string stating the name of the clustering Algorithm used, KSeeds, Kmeans or PAM}
}
\value{
vectorAUC vector containing the various AUC values for the various folds. Moreover the function draw the graph of AUC
}
\description{
Function Implementing metrics calculation of AUC
}

