% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{AUPR}
\alias{AUPR}
\title{AUPR}
\usage{
AUPR(predizioni, testpharmat, vectorAUPR, name)
}
\arguments{
\item{predizioni}{matrix of predictions}

\item{testpharmat}{matrix of test for the Side Effects}

\item{vectorAUPR}{empty vector to store AUPR}

\item{name}{name of the clustering algorithm used (KSeeds, KMeans,PAM)}
}
\value{
vectorAUPR vector containing AUPR values for the various folds, the function also draws AUPR graphs
}
\description{
Function Implementing metrics calculation of AUPR
}
\examples{
#Function for obtaining AUC
#Once you have obtained predizioni with the Prediction function you can apply
#this AUPR function using the following command (testpharmat sideeffects test matrix)
#vectorAUPR<-numeric()
#vectorAUPR<-AUPR(predizioni,testpharmat,vectorAUPR,"KSeeds")
}

