% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{CreateFolds}
\alias{CreateFolds}
\title{CreateFolds}
\usage{
CreateFolds(features, num_folds)
}
\arguments{
\item{features}{is the features matrix that has to be divided in folds for performing cross validation}

\item{num_folds}{number of folds desired}
}
\value{
folds: the elements divided in folds
}
\description{
Create the folds given the features matrix
}

