% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugClustKMeans.R
\name{DrugClustKMeans}
\alias{DrugClustKMeans}
\title{DrugClustKMeans}
\usage{
DrugClustKMeans(num_folds, num_clusters, num_iterations, features, side_effects)
}
\arguments{
\item{num_folds}{number of folds}

\item{num_clusters}{number of clusters}

\item{num_iterations}{number of iterations}

\item{features}{features matrix}

\item{side_effects}{side_effects matrix}
}
\description{
Function Implementing DrugClust with KMeans algorithm
}

