% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugClustKMeans.R
\name{DrugClustKMeansEnrichment}
\alias{DrugClustKMeansEnrichment}
\title{DrugClustKMeansEnrichment}
\usage{
DrugClustKMeansEnrichment(num_clusters, features, pharmat)
}
\arguments{
\item{num_clusters}{number of clusters desired}

\item{features}{matrix features}

\item{pharmat}{matrix of side effects}
}
\description{
Function Implementing DrugClust with KMeans and Enrichment
}

