% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugClustKSeeds.R
\name{DrugClustKSeedsEnrichment}
\alias{DrugClustKSeedsEnrichment}
\title{DrugClustKSeedsEnrichment}
\usage{
DrugClustKSeedsEnrichment(num_clusters, features, pharmat)
}
\arguments{
\item{num_clusters}{number of clusters}

\item{features}{matrix of features}

\item{pharmat}{matrix of side effects}
}
\value{
number of pathways for various clusters
}
\description{
Function Implementing DrugClust with KSeeds and Enrichment
}
\examples{
#features is the features matrix
#resultSeeds<-DrugClustKSeedsEnrichment(4,features)
}

