% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DrugClustPAM.R
\name{DrugClustPAMEnrichment}
\alias{DrugClustPAMEnrichment}
\title{DrugClustPAMEnrichment}
\usage{
DrugClustPAMEnrichment(num_clusters, features)
}
\arguments{
\item{num_clusters}{number of clusters desired}

\item{features}{matrix of features}
}
\value{
number of pathways for various clusters
}
\description{
Function Implementing DrugClust with PAM algorithm and Enrichment
}
\examples{
#features is the features matrix
#resultSeeds<-DrugClustPAMEnrichment(4,features)
}

