% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{Enrichment_Proteins}
\alias{Enrichment_Proteins}
\title{Enrichment_Proteins}
\usage{
Enrichment_Proteins(features, num_clusters, clusters)
}
\arguments{
\item{features}{matrix of features}

\item{num_clusters}{number of clusters}

\item{clusters}{clusters returned from the clustering algorithms}
}
\value{
vector_numb_pathway return a vector telling in how many pathways the various clusters are involved
}
\description{
Function Performing Proteins Enrichment using Gene Ontology
}
\examples{
#feature is the feature matrix
# pamx is the result of the PAM function
# and pamx$clustering gives the list assigning each element to a certain cluster
#all_pathways<-Enrichment_Proteins(features,4,pamx$clustering)
}

