% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{InitSideEffect}
\alias{InitSideEffect}
\alias{InitSideEffects}
\title{InitSideEffect}
\usage{
InitSideEffects(nameSideEffects)
}
\arguments{
\item{nameSideEffects}{name of the file where the side effects are stored. The format has to be a binary matrix, where the rows are the drugs and columns are the various side effects (1/0 meaning presence or absence of a certain side effect).}
}
\value{
The matrix containing drugs side effects
}
\description{
Initialize the matrix of features and Side Effects
}

