% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{KMeansModel}
\alias{KMeansModel}
\title{KMeansModel}
\usage{
KMeansModel(train, trainpharmat, num_clusters, cl)
}
\arguments{
\item{train}{matrix of train features}

\item{trainpharmat}{matrix of training of side_effects}

\item{num_clusters}{number of clusters desired}

\item{cl}{results of the KMeans model clustering function}
}
\value{
A Bayesian matrix of model for predictions, given the KMeans clustering
}
\description{
Function finding the Bayesian Model given the KMeans clustering algorithm
}

