% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{KSeedsScores}
\alias{KSeedsScores}
\title{KSeedsScores}
\usage{
KSeedsScores(train, trainpharmat, num_clusters, Seed, s, clusters)
}
\arguments{
\item{train}{train matrix of features}

\item{trainpharmat}{train matrix of side effects}

\item{num_clusters}{number of clusters desired}

\item{Seed}{subset of the features matrix containing only the Seeds drugs}

\item{s}{the seeds of the clusters}

\item{clusters}{the list of clusters where the various drugs are}
}
\value{
A matrix containing prediction scores for each cluster
}
\description{
Function for obtaining the Bayesian prediction scores using KSeeds clustering
}

