% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{PAM}
\alias{PAM}
\title{PAM}
\usage{
PAM(train, num_clusters)
}
\arguments{
\item{train}{matrix of train features}

\item{num_clusters}{number of clusters desired}
}
\value{
pamx structure with various values resulting from PAM clustering algorithm
}
\description{
PAM clustering algorithm
}
\examples{
#train is the train feature matrix
#pamx<-PAM(train,4)
}

