% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{PAM_Model}
\alias{PAM_Model}
\title{PAM_Model}
\usage{
PAM_Model(pamx, num_clusters, trainpharmat, train)
}
\arguments{
\item{pamx}{result of pam clustering algorithm}

\item{num_clusters}{number of clusters desired}

\item{trainpharmat}{matrix of training for side effects}

\item{train}{matrix of train features}
}
\value{
A matrix of model for prediction of uncharacterised drugs, given PAM clustering
}
\description{
PAM clustering algorithm Model
}

