% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{Prediction_PAM}
\alias{PredictionPAM}
\alias{Prediction_PAM}
\title{Prediction_PAM}
\usage{
PredictionPAM(A, pamx, test, numb_sideEffects)
}
\arguments{
\item{A}{prediction scores matrix}

\item{pamx}{result of pam clustering algorithm}

\item{test}{test features matrix}

\item{numb_sideEffects}{number of side effects}
}
\value{
predizioni matrix of predictions given PAM clustering
}
\description{
PAM prediction models
}

