% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{PredictionKMeans}
\alias{PredictionKMeans}
\title{PredictionKMeans}
\usage{
PredictionKMeans(A, cl, test)
}
\arguments{
\item{A}{Bayesian model given by the application of KMeansModel algorithm}

\item{cl}{structure of clusters given by the KMeans function}

\item{test}{test matrix of drugs}
}
\value{
predizioni matrix with a number of rows equal to the number of clusters and a number of columns equal to the features
}
\description{
Function finding the predictions for the uncharacterized drugs given the KMeans clustering algorithm
}

