% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{PredictionKSeeds}
\alias{PredictionKSeeds}
\title{PredictionKSeeds}
\usage{
PredictionKSeeds(test, Seed, num_clusters, A)
}
\arguments{
\item{test}{test drugs features matrix}

\item{Seed}{matrix of seeds initialize in the KSeed algorithm}

\item{num_clusters}{number of clusters desired}

\item{A}{matrix of Naive Bayes predictions scores, result of KSeedsScores function}
}
\value{
predizioni matrix containing predictions for the various uncharacterized drugs
}
\description{
Function implementing predictions for uncharacterized drugs
}

