% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{RandomSeedGenerator}
\alias{RandomSeedGenerator}
\title{RandomSeedGenerator}
\usage{
RandomSeedGenerator(num_clusters, numbrowfeatures)
}
\arguments{
\item{num_clusters}{number of clusters desired}

\item{numbrowfeatures}{number of rows of the features matrix}
}
\value{
s list of seeds
}
\description{
Initizalize seeds for the KSeeds clustering algorithm
}
\examples{
r <- 8
c <- 10
m0 <- matrix(0, r, c)
num_clusters=4
features<-apply(m0, c(1,2), function(x) sample(c(0,1),1))
s<-RandomSeedGenerator(4,nrow(features))
}

