% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{InitFeatures}
\alias{InitFeatures}
\title{InitFeatures}
\usage{
InitFeatures(namefeatures)
}
\arguments{
\item{namefeatures}{name of the file where the features are stored.}
}
\value{
The matrix containing drugs features
}
\description{
Initialize the features matrix. The data needs to be binary matrices where each row is a
drug, and columns represents drugs features. If the element in position ij is 1 it means that the ith drug interacts with the jth element
(for example a protein). The same for the matrix where side effects are stored.
}

