% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{KMeansClusteringAlgorithm}
\alias{KMeans}
\alias{KMeansClusteringAlgorithm}
\title{KMeans}
\usage{
KMeans(train, num_clusters)
}
\arguments{
\item{train}{matrix of train features}

\item{num_clusters}{number of clusters desired}
}
\value{
cl list containing the clusters ownerships
}
\description{
KMeans clustering algorithm
}
\examples{
#use the initFeatures to upload train feature matrix
#see also KSeedsClusters function to see a similar example
#with a toy matrix
#cl<-KMeans(train,num_clusters)
}

