% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/VariousFunctions.R
\name{RandomSeedGenerator}
\alias{RandomSeedGenerator}
\title{RandomSeedGenerator}
\usage{
RandomSeedGenerator(num_clusters)
}
\arguments{
\item{num_clusters}{number of clusters desired}
}
\value{
s list of seeds
}
\description{
Initizalize seeds for the KSeeds clustering algorithm
}
\examples{
s<-RandomSeedGenerator(4)
}

